refactor(ocr): 将Tesseract服务文件移动到tesseract子目录并更新配置
